Add stack canary to detect overflows.
Placed the task structures and task pointers/scheduler thread
in seperate sections. With the structs before everything else.
This should prevent corruption of current_thread on overflow.

This catches the previous AArch64/O3/UBSAN failure, but only
with a stack size of 512. At 1024 I beleive we have enough to
get past the stack check, but then saving registers overflows
the stack.

Checking we have enough space to save our context is an area
for further work.