Fix UBSAN type mismatch on AArch64 (all opt levels)
Adding the stack canary appears to have disturbed some
code which was sensitive to stack alignment on AAch64.

get_msg was failing with a type mismatch and my working
theory is that the code generated a pointer assuming 16
byte stack alignment, but the underlying data was actually
a different type.

To fix this, align the stack pointer to 16 for all platforms
when we init the threads. This means that even if you make
an unaligned Thread struct, the stack pointer will end up
aligned anyway. (and we don't have to mess with packing)