Fix qemu_print infinitley looping.
A hlt instruction is a software breakpoint.
So it's exception link address is not the *next*
instruction but the hlt instruction itself.

This meant that qemu did the semihosting print,
then returned control back to the exception handler,
which eret back to the thread's hlt instruction.

Causing an inifite loop. To fix this eret to one
instruction beyond the hlt. Really we need to use
svc with different codes for this.