Use svc to ask monitor to do semihosting instead of directly using hlt
Now all exceptions will be svc calls with either 0xdead for thread
switches and 0x3333 for semihosting.

Fixed an issue where stack underflow was not caught when it was
happening in the scheduler thread. The scheduler thread was in the
same section as the other important globals when it should have been
with the other threads so an underflow couldn't overwrite crucial data.

It now lives with the other threads and stack underfow is correctly
reported.