Bring all the ports in line with the startup.s/vector.s format
Now all ports have a vector.s which is placed at 0,
and a startup.s which holds the reset code.

AArch64 is a bit of a cheat, it just has an empty vectors
file. This is due to not being able to branch from 0
to the start of RAM. AArch64 simply configures its own
vector location as it did before.