On Arm do user stack extent check using the monitor stack.
We use supervisor mode's stack pointer to save
a couple of temporary regiters to allow us to decide
whether it's a thread switch or a semihosting call.

Then for semihosting calls we switch to system mode,
which shares the sp of user.
For thread switches we stay in supervisor and check
the user stack extent. By temporarily switching to system
and grabing user's stack pointer.

Then once we've done that restore the temporary registers
from the monitor stack and switch to user mode to do the switch.

Eventually we should do the switch in system mode,
so we can save all registers not just callee saved.