Save all registers and CPSR in Arm thread switch.
Use supervisor mode to verify monitor memory
and check user stack extent. Then move to system
mode to do the save restore.

The initial switch into the scheduler originally
needed multiple entry points but this is difficult
considering which modes you'll be in.

Instead of this add a small asm routine to setup
the stack pointer of the dummy thread normally
so that it passes the stack check without any
skipping ahead.