Add stack overflow canary.
This means we can detect when someone uses more stack space
than is allocated. As in writes off the end of the stack
the compiler gives them, enough to hit the top of the thread stack.

This is on top of the existing stack underflow detection.
Which covers the event that someone has deeply nested calls
that decrement the stack pointer too far.
(assuming they actually write to memory there, which is likely)

Tests to follow.