Reduce entry memory by putting dummy thread in globals.
This way we don't have to adjust the stack allocated in
the linker script when adjusting the thread's stack space.

Use log_event for stack underflow error so we can see it
in the logs. Except that we'll always null the name in case
it was set to some random unterminated memory.