Misc. improvements on the way to a test for stack checking
Add an option to configure whether we exit qemu or invalidate
the thread and try to contiune when there's a stack error.

NULL the name and set ID to -1 before logging anything to make
sure we don't print invaid data.

If we aren't going to exit, use the dummy thread's stack to
be able to yield back to the scheduler. (after marking the
current thread as invalid)

In the scheduler check that the index in the all_threads array
we're checking either has an invalid id (-1) or one that matches
the index. If this check fails it could indicate undetected corruption
or an error in our stack handling.