Don't call semihosting functions on Thumb when we get an invalid svc
After much wrangling I've decided not to bother printing an error
here. Just branch to self and you can stop it in GDB easily.

The alternative was a bunch of macros to tell whether we were
calling semihosting from privleged mode (needs a bkpt) or not
(svc instruction). This would't just be thumb, all the ports would
need it. That was more assembly than I can be bothered to maintain.