Save all registers on Thumb, not some.
Thumb wasn't saving the high registers, just r4-7.
This caused problems in the stackcheck demo at O3.

When print_thread_id was inlined some register was being
used to hold a value/pointer to the thread name it should
print.

I beleive that when this sequence was interrupted by yield()
that register was trashed or not updated and that lead to it
printing an empty name after the overflow/underflow threads
had run.