Add the ability to pass arguments to threads.
Each thread can take up to 4 arguments of up to
their register size (sizeof(void*). They are passed
to the worker function by thread_start.

A macro make_args is provided to automatically do
the casts. It assumes 4 arguments, we could add 4
versions of it but easier to just pass 0 to unused
for now.