Don't misdiagnose Thumb instructions as SVCs in exception handler
We first checked whether the instruction ended in 0xAB or 0xFF
which made sense when they were the only reason for exceptions.

When the timer interrupt was added, what could happen (and did
on O0 with UBSAN) is an interrupt on an instruction that wasn't
an SVC. That also ended in one of our magic numbers.

To fix this we check the exception number first, to decide
between timer interrupt or SVC. If it's none of those just spin
for debugging later. Handle the timer interrupt as before and
if it was an svc instead, then go read the instruction from memory.