Got the timer interrupt firing on Arm.
Thanks in no small part to: https://github.com/littlekernel/lk/issues/54

Otherwise the GIC config is exactly the same as AArch64 (GIC v3).

On the way fixed the way we were returning from exceptions
when calling semihosting functions. (and the new timer svcs)
The way we were doing it must have corrupted sp or fp, because
local variables on the stack were different before and after
the semihosting calls.

Interesting to note that UBSAN caught this at O0. It's usually
the higher opt levels causing issues, but like I said, it was about
stack use and O0 has plenty of that.