Implement timer handling for Arm.
We have a seperate IRQ handler like on AArch64.
Then we do a switcheroo to get the IRQ mode's lr
(after correction) into supervisor mode's lr.

From there we act like it's a normal SVC handler.
Except that we set next thread to the scheduler
and mask further timer interrupts along the way.