Fix race condition in the timer demo, found when adding Arm support.
If the timer value is long enough, it won't fire between enable
and disable. This was happening on AArch64 and Thumb and meant
that the threads exited immediately after total_prints became 0.

On Arm, I happened to choose a smaller value for the timer. This
meant that sometimes it would act as the others did and sometimes
it would need to be scheduled a second time to exit.

To fix this I have updated the the demo to expect the behaviour
seen on Arm, but consistently across all 3 platforms.