Dedupe the GIC setup code.
Arm and AArch64 use the virt board so all the locations
are the same. Had to break out the ifdefs to get a valid
return instruction for Armv7a.

Also note that gic_init is only a naked function on Armv7a.
aarch64-elf ignores that attribute. This is fine though
because AArch64 startup will set a valid stack pointer
before calling it.