Fix error on Arm O3 when calling gic init.
Lower than O3, we could get away with a bl to
gic_init and using the lr to return from it.

At O3 the compiler starts to use lr as a general
register. Which makes sense considering we've told
it that it's a naked function.

This means we can't return using the value in lr.
To work around this, asign a global label to the
point just after the call to gic_init and branch
directly to it.