Add emulated TLS for Arm and Thumb.
This provides __emutls_get_address which GCC will call
to when you use '__thread'. I'm statically allocating the
address arrays with one per thread and a max number of
thread local vars as defines.

To allocate the variables I've got a simple buffer with a
fixed size.

AArch64 not supported yet, it uses the tpidr_el0. So I
either support that, or find a way to force it to use
emulated TLS.