Switch to Linaro toolchain for arm bare metal.
Had to diable the build ID section as it was moving the
vector table.

Also TLS must either be implemented fully or emulation
was disabled in this version of GCC. (probably the former)
So I've had to disable the emutls demo on Arm as well,
since I have no way to force the use of emutls.