Use a thread state instead of storing the current PC in the task struct.
This means we can add more in between states later.
For now we have:
- init, never run before
- suspended, has yielded
- running, is the current thread
- finished, has exited thread_start

So we still use an id of -1 to indicate an invalid
thread struct in the array, but now we have an additional
check of the state.

Basically if the id of the thread is not -1, then we check
if the state is suspended. If it is, it can be scheduled.
(and from a user program point of view, it is only 'valid'
under those same conditions, see the spawn demo)