Set the states in assembler to avoid race conditions.
E.g if you do:
  set_thread_state(suspended);
  --> interrupt makes you yield
  <-- return from interrupt with state=running
  thread_switch()

Then state is still running, but you've yielded.

How to solve this for thread's finishing is an
open problem. For now I've stuck to invalidating
the thread id. Though this is really only meant to
mark the thread struct as invalid, and a finished
thread still has a valid struct.