AArch64 asm changes for thread states.
I also realised that we're not returning from switches
how I thought we were. We're returning to what the link
register was when we called thread switch.

Meaning we don't return to thread_switch, we return to
the thing that *called* thread_switch. (usually thread_yield)

Which is fine, but was a bit mind bending.