Properly set finished state.
Set finished state before exiting the thread.
thread_switch will now only set the state to
suspended if it was previously running.

Add a few more utility functions that can
be used to handle the difference between
a thread id of -1 meaning exited badly,
and a state of finished with a valid thread id.