Add condition variables and new demo for them.
This is implemented with a new 'waiting' state
for threads. This state will be ignored by all
scheduling calls like yield/yield_next etc.
(yield_to will not let you yield to a waiting
thread either)

When the thread is signalled via the condition
variable, its state will become suspended. Then
it'll be run the next time it would normally be.