Add thread_cancel feature and demo.
Along the way refactor some of the scheduling stuff and
update demos accordingly.

yield_next will now just return if the calling thread
is the only active thread, instead of switching to itself.

Return the thread state from thread_join so you can tell
if it was destroyed (returns false) or finished or cancelled.
Also allows the stackcheck demo to use it.