Add scripts to generate and run random demos.
The basic overview is:
  * add a number of intial threads. Amount randomly
    chosen from 0 to some set maximum.
  * generate a worker function for each thread,
    with a number of 'actions' (aka API calls).
    Again randomly chosen from 0 to some max.
  * For any threads this generated worker creates,
    repeat the process. This can continue to some max
    depth of add_thread->worker->add_thread->worker etc.
  * print out the C code for this and put it in demos/generated.c
  * run it and see what happens.

There's a script to invoke all that:
./run_random_demos.sh

Which will also check the formatting of log messages.
(mainly an issue when the timer interrupt is enabled)

We expect all generated demos to finish with all threads
finished or cancelled. This is because I've avoided bits of
the API that could lead to infinite loops.

Having said that enabling timer functions causes a lot of that
so clearly there's issues there.

Fixed a small bug found in initial runs. add_thread would go
out of bounds if it didn't find a free space in the thread array.
Apart from that nothing much turned up. (without interrupts)