Fix randomly generated demos with timers on AArch64.
We were only disabling the timer when we
got an interrupt from it. This meant that
when you yielded normally it would be left on.
Meaning you could interrupt the scheduler.

This wouldn't be an issue but for the comment
at the bottom of this file. If we skip a stack frame
in the scheduler we could choose an invalid thread.

I need to confirm the logic of the comment to see
what the correct soloution is. I'm not even sure
it was correct to begin with.