Revert "Fix randomly generated demos with timers on AArch64."
This reverts commit 911fcc4d234d56b1753ccfacf9d9fe39b697b383.

While this does fix the randomly generated demos it misses
the point. AArch64 doesn't return from interrupts correctly.

If it did, interrupting the scheduler wouldn't be an issue.
It'd simply switch to itself and carry on.

Appears to not affect AArch64 Linux but I think that's just
a happy coincidence.