Correctly save exception link register on AArch64
What we should have been/are now doing is saving
ELR_EL1 onto the stack like Arm/Thumb were doing
for us.

* Save ELR_EL1 onto stack when a thread yields
* Place thread start address in ELR_EL1 if it's
  a first run
* adjust the required stack size by +8 for the pc
* restore ELR_EL1 from stack before doing the eret

This properly fixes the selfyield demo on AArch64.

Along the way fix the stack extent check. It should
have been x8 for bytes not x64 for bits, and the
check was the wrong way around.
(why it worked at all I'm not sure)