Automatically call start scheduler
The last thing we always do is start the
scheduler so just do it in the kernel.
That way we can't forget it.

Rename the first user function to "setup"
to reflect this.

Updated all the demos. Only the args demo
was relying on the stack frame of demo()
essentially being static.
So those strings are now static globals.