Handle printing thread IDs > 9
Previously we assumed a thread ID would only
be one digit. So move the uint_to_str that UBSAN
was using into print.c and share it.

Return the number of digits in the resulting
string so we can align it in the log output.

Added a demo (a test really) for this behaviour.
Aligning a single/double and named thread in output.

I'm going to stop documenting the more "test" kinda
demos. In future it's probably better to seperate them
into their own folder.