Add variadic print with string formatting
This replaces all the repeated calls to print
with single strings. Now we can use %s to insert
strings like printf does, with the usual %% to
escape an actual %.

Meaning that thread names and UBSAN output are
now one print call with many %s.

Anything other than %s will be an error and caught
by UBSAN if it's on.