Add sprintf with %u for thread names.
This means we can use sprintf on Linux with the
same code in thread.c.

Along the way fixed a bug in printf which would
print the null terminator if you ended your
string with a format specifier. This wasn't visible
for printf but was a hard error for sprintf.