Correct the multiple !! before OCLint sees them in ASSERT macro.
Use a new scope so I don't have to work out how to concat
file and line properly for a unique var name.

Can't use the usual //!OCLINT here since we're in a macro.
So declare a temporary bool which will resolve the first levels
of !/whatever then just check whether !condition is true.

This means there's no issue for OClint to pick up.