Add implementation of malloc/free.
Fixed block size heap with no reordering
or moving. If it's too fragmented malloc
will just fail.

Though alloc.h does have an include for
Linux, the demo won't run on Linux. This
is because it's relying on properties of
the implementation here.

However we can now write other demos that
use malloc/free on bare metal or Linux.

Needs a mutex adding for it to be thread safe.