Implement a PThread layer for running on Linux.
This removes the need to use any assembly,
and can run on anything with PThreads.

To make the scheduling deterministic we
essentially keep yieling until we get
the pthread we think should be the one
running.

Next thread was made volatile to prevent
LTO re-ordering writes to it in this
configuration.

Lots of ifdefs in thread.c for now,
could do with some cleanup.

This also means we can run the TLS demo!
Since the underlying mechanism now supports
TLS out of the box.