Improve realloc by freeing first.
Now that free doesn't actually touch the
data we can free the current allocation,
then look for a new one. This will include
the current allocation as free space.

Then if it fails, simply restore the tag
to the current allocation and nothing has
changed.

Before we couldn't do this because the tags
were in the heap itself. So free was actually
overwriting what was user data.