Add a basic in memory file system.
Up the heap size so we can run the tests for
it on AArch64.

The filesystem demo is in addition to the
existing file demo. When building the former
we'll use the functions from file_system.c instead
of the semihosting ones. Exit will always be done by
semihosting.

This means that we have these configs now:
Linux      - posix       / in memory
Bare metal - semihosting / in memory

Stil some ASAN failures on Linux to fix up.