Don't build with Ninja
It's not so much building that's the issue.
It's that I run tests via lit which invokes
ninja on multiple threads. These copies fight
for the log file hence the inconsistent errors.

I could run lit with -j 1 but that's just another
bad hack. Really we'd not be using the build
system to run tests.

Keep 18.04 image since it matches my local box.