Remove the need for thread_switch_initial on Arm
By making the svc handler only save the current
state (and hence check stack extent) if the current
thread has run.

This means that we can directly switch into the
scheduler thread as it'll be at init.
Also using the same trick when we get a stack
check failure.
(though that's inviting a race condition but
it was already doing that so one more on the pile)