Fix UBSAN assert on Thumb O3 LTO
This was because we were actually using
more than the 64k of RAM the thumb board
provides.

To solve this move the text section to
before RAM and add an assert after BSS
so we'll catch this in future at build time.

Update the local test script to match CI
while I'm here.