Remove mlong-calls, fix the fallout
mlong-calls was our poor man's PIE for
a bit. Since we've given up on that for
now we know where the bin will be.

However, thumb fns will be 4 byte aligned
so we need to force the loaded_code array to
also be.

It wasn't and this meant all the jumps to veneers
were 2 bytes off in the second worker function.