Address Arm semihosting not returning in kernel mode
In a story that will be familiar, I found that handling
an svc in supervisor mode is possible but confusing.

Given that we already took the easy way out for Thumb,
do the same for Arm and check cpsr to see if we're in the
kernel.

Add a call in the file demo's setup to check that we
actually come back from semihosting.