Fix stack corruption on resume on AArch64
The reason we had this issue was that
when the first loaded thread switched
for the first time it was in *EL1*.

Which is unexpected to say the least.
So stack pointers got confused, we restored
what was last used for EL0 which happened
to point us back to the start_thread.
(presumably because the thread we added
*did* start in EL0 and set this up)

So just make sure that when we fake the return
PC for EL0 also set SPSR so we're actually going
back to EL0 as we expect.