Don't run loadbinary with LTO on
When we have lto code_page is removed as unused.
Which is true to some extent.

We can solve this by sprinkling attribute used
*everywhere*. However that's just one issue around
the kernel's ABI.
For example you can't build a loadable at -O3 for
a kernel at -O1 because the addreses change.

This could all be solved by moving to syscalls for
everything but I'm not about to do that now.

Short term fix, don't run it with LTO and we'll always
be matching compiler settings for non LTO builds
anyway.