Allow loading a new binary if the first one has finished
The lifetime of the current code page is set as
the lifetime of the initial loaded thread, and any
threads that thread creates. (and so on and so on)

Use of the static in worker proves that we do reload
this. Also a thread in the kernel file tries to reload
while some threads are still using the code page, to
prove that the check is working.