Add support for loading multiple binaries
By using backing memory to the existing
active code page. For now we're not counting
the main page in the total so if CODE_BACKING_PAGES
is N, you can load N binaries.

Until one of them finishes all it's threads
then the page is freed for another load to use.

To handle situations where you'd normally skip
the scheduler I've moved the NULL next_thread
check into C in do_scheduler.

This means that if you yield_to etc, then
it'll still go to the scheduler and do any
needed swapping.
(and it means user mode isn't doing them,
which is the point really)

Things checked by the test:
* a thread in a backing page is copied
  to the code page before running
* a thread suspended is copied back to
  the backing page
* Doing yield_to or similair directed
  switches doesn't change this
* And when you do this the code is different
  if expected to be (hence task2.c)
* You can't load more programs than you
  have backing pages
* The lifetime of the backing page is
  that of any threads created from it
* code and data is moved when we're
  paging