Handle relocations to symbols in the ELF itself
If the section number of a symbol is not UND then
it refers to something in the current file.
We then offset that to get the final runtime address.

Fix logging of symbol bind values.

Change "address" to "value" in a lot of places since
the value can be, but is not always an address.