Don't run loadpiebinary demo with sanitisers
This would need us to add all the handlers to
the kernel symbol table. Which is a manual thing
at the moment. Better do that later when there's
a better way.

Added a bit of undefined behaviour to the pie binary
code to catch this if we accidentially do run it
with sanitisers.