Support PIE for AArch64
New relocation types which are basically the
same as Arm except they are relA this time.
So there is a signed offset to consider.

Instead of two code paths I've just ifdef'd
the addend field and asserted that we only
see relA on AArch64 and plain rel elsewhere.

So the addend *should* work however the demo
only ever generates zero in that field so it's
not really proven.

Modified the kernel linker script to align the
code page to 4k because this is required to use
adrp in AArch64.

Seems to be ok with the other ports, I was concerned
we might run out of space.
(maybe you do on the backing page demo?)