Yielding officially the first demo to work with syscall ABI
For syscalls like yield() that cause a thread switch but return,
run the scheduler to set next thread.
If we see it set after coming back from C to assembly handler,
jump to thread switching instead of returning normally.
We'll patch the return value of the syscall back into the
saved registers of the current thread so it thinks it just
returned as normal.