Exyielding demo mostly working
Only problem is that we log yielding/resuming
even if we don't yield_to/yield_next anything.

We could do something where you check if the target
thread is valid/there are other threads. Then ask to switch.
But considering the whole point of this was to isolate
this kinda multi part call into single syscalls that don't get
interrupted, I don't think that's a good idea.

So just update the expected test output.