Update threadlocalstorage demo
Fixed an issue where the scheduler would exit
if you yielded from the last thread. This must
have changed recently, not sure how it ever worked.

Also switched to user calls in emutls. Though this
doesn't make much difference now. I think we'll need
to come back to it if it's to be thread safe eventualy.