Update loadbinary demo
For this we need to handle loading thread
from a file being a syscall but not always
being present.

Always put it in the enum and if it's not
enabled simply replace it's place in the
handler table with a branch to self.

Not the greatest error handling but we'll
deal with that later.

Also move the handler table into syscall.h
and do the same trick we did with thread_state.h.
When included in C you get the normal stuff,
when in assembler you get the table.