Update loadbinaries demo
do_scheduler was being run unconditonally before
loading the next thread. This caused us to swap
code pages twice.

So revert to the old behaviuour. If we get to
loading the next thread and next_thread is NULL,
run the scheduler. Otherwise assume that it has
already run and done all the housekeeping we need.

Along the way realised that yield_to and yield_next
actually skip stack checks. So have those call
k_thread_yield so we don't miss that.