Fix O1 LTO run
When we did the first thread switch we didn't
give up the kernel's inital stack frames.

Somehow (thanks to LTO I guess) this led to us
calling do_scheduler from assembly and picking up
some of this old data.

Again not sure how but the upshot was us overwriting
the stack pointer of thread 0 with an address in ROM.

So init the kernel stack each time we call do_scheduler
from assembly.