Don't reset stack before do_scheduler
I think this was simply masking some other
issue that got fixed later.

It is true that because we call start_thread_switch
directly from entry() we are giving up a stack frame there.
Since entry() never returns to reset it.
However we shouldn't reference that frame so it's not an
issue.