Missed r7 as a clobber register in syscall invoker
Really this clobbers all the registers but we
can get away with this and it's neater for now.
Caller saved is handled by the kernel and we
assume that you aren't doing anything after the SVC
but returning r0 so the rest of the regs being
changed isn't an issue.
(until it is, cross that when we come to it)