Fix syscalling on AArch64
Originally it was using x8 to match Arm's r8.
On Arm 8 is callee saved and was marked as a clobber.
On AArch64 x8 is caller saved so despite the clobber
I copied it to a callee saved register to make use of
it in the syscall handler.

This register was not on the clobber list.

So to fix this, simply switch to the callee saved x20
register for AArch64 syscalls.