Remove the Linux port
It's a pain to convert to emulating the syscalls
for a start. Going forward there's only going
to be more of that kind of issue.

We lose coverage and Clang compiles but the latter
at least we could probably do with the correct
sysroot settings if we wanted it back.