Reduce add thread syscalls to just one
This will take a name, args, worker and a kind.
Where kind is THREAD_FUNC for functions or
THREAD_FILE for loading elfs.

Additionally, set worker to be the path to the
elf when doing so.

The permutations are handled on the user side
with some forwarding functions and on the kernel
side with macros.