"fix" stackcheck failure on Arm Os UBSAN LTO
When we asked for the thread name in log_event
we didn't init the ptr before the syscall.
After fixing that a lot of things failed and
printed only IDs.

Seems like I'm missing a volatile somewhere,
but for now a memory barrier after getting the name
seems to do the trick. Pending some better
thought out soloution.
(I have a suspicion that this could affect other
calls)

Changed the stackcheck demo to use set name API
call which needs a correct thread ID so patch
that in after underflowing.