Able to write PC via regset
Caveats:
* threads that come in through yield() syscall
  haven't saved data until that ends, so beware of that.
  (shoudn't be an issue from user mode)
* using the regset on an init thread is bad since that
  has no context to work with yet

On the plus side it works, we can redirect the thread
about to be run.

I'm dealing with the stack ptr being in the context
not on the stack by putting that on the end of the struct.
Then writing everyting but that back to the stack.
(also writing to sp will break a lot of stuff since
it's cruical for restoring threads)