Fix arg structs not being set
For some reason recent changes caused gcc to
realise that no one reads the ThreadArgs passed
to add_thread.

So work around this with a volatile assignment in
user/add_thread. This applies to any struct passed
by pointer so maybe adding volatile to all the
interfaces is better, but I'm not going to do that now.

(perhaps a thread name set from a stack variable could
also have the same issue)