Add error for elf file not found
Although it is masking the other kind
of error, when the file isn't the right
format. Let's ignore that for now.

Bigger issue is that this has exposed that
setup() has an id of -1 therefore you can't
get any of its properties.

I think making setup thread 0 is worth considering.
It saves some memory space and it continues
the thread id == index in array assumption.
Even if I do have to adjust all the demos.
(would also account for weird stuff like yielding
out of setup)