Rough cut of signal handler
We do this by pushing a fake saved context to the
thread's stack, resuming into a jump to the handler
and detecting the resume address being the end of that
jump function.

Only working for thumb. Issues:
* there really should be permisson for setting your
  own attributes vs setting other thread attrs
* providing the stack pointer in register context
  is:
  A: causing errors (sizeof(RegisterContext) isn't
  what we wanted here)
  B: misleading since if you modify the stack ptr
  you can't restore the thread properly anyway
* we have a single signal pending, perhaps the handler
  should get an argument of the type of signal?
* signals to a thread that hasn't run at least once
  are ignored. Not sure if this is annoying enough
  to write some special case code for.

That final point happens because init threads don't
restore anything (there's nothing *to* restore).
We could always put a blank context there.
Though it wouldn't help this use case as there's
no original context to use after the handler finishes.

Idea: add initial return context for threads in init_thread
instead of building it in assembler. Then special case state
init but going into a handler.