Support multiple signals
Pending signals is now a bitmask.
(caveat: multiple of the same value treated as one)

Also, setup the intial register context
in init_thread instead of in assembly.
This means we can treat init state threads
the same as any other.
(though most of the time doing things to an init
thread won't do much, but at least it won't break
them)