Further explain self backtrace comment
Also add a fallback for when you don't have
a symbols list.

Even having print_backtrace do the register
read for you has the same issue. It's always
going to be the frames at the point you asked
for the registers.

One way around this would be to have the kernel
do it.