Add fibres aka user threads
Inspired by:
https://graphitemaster.github.io/fibers/

This provides two main ways to use fibres.
The first being simple save/restore in the
same function.

get_context()
<restart point for the context> <--\
<some stuff>                       |
set_context(<earlier context>) ----/

The second is for swapping between fibres.
Where you make a new context that has its
own stack and starts in another function.

Then you swap_context to that context,
saving the current context along the way.

This other function should take a single
argument which is the context it would
swap back to.

void other(UserContext* ctx) {
  <stuff>
  swap_context(ctx);
  <more_stuff>
}

void use_fibre(void) {
  <...>
  make_context(&ctx, other, <top of stack for other>);

  <do work>
  swap_context(&ctx);
  <do more work>
  swap_context(&ctx);
}

When a fibre function returns normally
it will go back through set_context so that
you will always return to the other context.

This is done by setting the new context's lr
to set_context_from_stack_address and putting
ctx onto the new fibre's stack.
(with some padding on AArch64 for 16 byte alignment,
using the added ALIGN_STACK_PTR port macro)

When it returns we pop that pointer and use
it to switch back to the original thread.

Really the new assembly should go in some port
folder in src/user but it can stay in the kernel
asm for now.