Properly copy/init data/bss from ROM...
...instead of loading directly into RAM.

This fixes #1 where QEMU 5.1.0 was giving
an error because we were using the -kernel
argument but loading data into RAM.

I think this led to a conflict when QEMU
reset the board, as it tried to init the RAM
where we had copied our code.

Earlier versions of QEMU (at least 2.x) were
less strict about this.

To fix this we:
* place initialised data into ROM but with
  a runtime address in RAM
* add linker vars for start/end of data and
  bss sections
* copy/init those in kernel entry()

See:
https://sourceware.org/binutils/docs/ld/Output-Section-LMA.html#Output-Section-LMA

Then use "-device loader,file=<kernel>" instead
of "-kernel". This is the more "bare metal" option.

Note: for Thumb older versions of QEMU (as we have
in CI) *required* -kernel. So we'll use it there
for now and switch Thumb over once CI moves to
Ubuntu 20.04.

Doing this exposed a stack overflow on AArch64
when loading the pie ELF. So I have upped the kernel
stack size to 3kb.
(maybe this should be configurable because...)

This meant that on Thumb, loadbinaries no longer
fits in 64k of RAM. Luckily QEMU doesn't seem to
check whether you use more than that so I've bumped
Thumb to 70k, which is pushing our luck but not by
much. (fingers crossed)

Final thing to note is that I have told the linker
to discard Arm unwinding information. We could put
that in the data section but it doesn't make much
difference.
(maybe we'll actually use it one day)

See:
https://sourceware.org/binutils/docs/ld/Output-Section-Discarding.html