Fix loading of .bss section
We were assuming that there were zeros
in the elf for us to copy. This is not
the case. There is an offset but it's
not supposed to be used.

So zero init any section that has
type SHT_NOBITS.

Added an assert in the loadbinary demo
to test for this.