Unify printf/sprintf handling
By using a sort of C class to hold the buffer
you're going to write to. Along with a "writer"
function that knows how to handle the buffer pointer.

For the serial port we don't move the buffer pointer,
for memory buffers we do.

You could do this passing around 2 arguments to each function
but this seems a bit easier.

Expanded the printthreadname demo to use printf and sprintf
for each formatting checked.