Use -fPIC when compiling PIE loadables
Fixes #2.

Current Linaro toolchain for AArch64 in CI is 7.4.1.

When compiling with toolchain from
https://developer.arm.com/tools-and-software/open-source-software/developer-tools/gnu-toolchain/gnu-a/downloads
which is 9.2.1 linking fails due to code incomopatible
with position independence being generated.

Add -fPIC to compile options so this does not happen.
I assume we got lucky on 7.4.1 and it just happened
to use a compatible code sequence.