Fix printthreadname demo at -Os
Though Os seems to not be the cause, just
the only setting where we saw the issue.

Which was that when we echoed back unknown
formatters we printed until the null terminator.
(this meant you saw the next test line printed
for each time the unknown line was run)

Which is not what you want to do if you have:
%? %0?

Since we only want to print to the next char,
in this case the space between ? and the second %.