[Thumb] Fix out of RAM error at -O3 building loadbinaries
With the recent print changes, the compiler realised
it could constant propagate and inline handle_format_char.
This added a lot of code size.

The propogated  copy was called by printf and sprintf
grew by around the same amount since it had another copy
inlined into it.

00006fa4 000010f0 t handle_format_char(int*, char const**,
std::__va_list, PrintOutput&) [clone .constprop.0]

00006fa4
<_ZL18handle_format_charPiPPKcSt9__va_listR11PrintOutput.constprop.0>:
    6fa4:       e92d 4ff0       stmdb   sp!, {r4, r5, r6, r7, r8, r9,
sl, fp, lr}
    6fa8:       b099            sub     sp, #100        ; 0x64

000081f0 <vprintf>:
    81f0: e92d 4ff0   stmdb sp!, {r4, r5, r6, r7, r8, r9, sl, fp, lr}
    <...>
    8262: f7fe fe9f   bl  6fa4 <<< calling handle_format_char

To "fix" this add attributes to only optimise handle_format_char to -O1
and disable inlining. (which probably doesn't happen at O1 anyway)

In some ways it's good that it picked this up with the refactor.
It just doesn't sit well with Thumb's memory limit and this
memory hungry demo.

Sizes with C print functions:
$ arm-none-eabi-size loadbinaries
   text    data     bss     dec     hex filename
  47241   18520   42912  108673   1a881 loadbinaries

With C++ print functions (including optimisation attributes):
$ arm-none-eabi-size loadbinaries
   text    data     bss     dec     hex filename
  46580   18372   43040  107992   1a5d8 loadbinaries

So in the end we've reduced the binary size. Albeit by only 0.6%,
with text down by 1.3% and data down by 0.79%.

As before, code size isn't a goal but being able to run
on Thumb is.